svg/painting/parsing/stroke-linejoin-computed.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: getComputedStyle().strokeLinejoin</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#StrokeLinejoinProperty" rel="help"/><h:meta content="stroke-linejoin computed value is as specified." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/computed-testcommon.js"/><script><![CDATA[

test_computed_value("stroke-linejoin", "miter");
test_computed_value("stroke-linejoin", "round");
test_computed_value("stroke-linejoin", "bevel");
// "miter-clip" and "arcs" are not yet supported by browsers.

  ]]></script></svg>
